Fixes (#70)
Reviewed a bit RenderToCube test.
Fence timeout is now an uint64_t (as should always have been).
Reworked memory barriers, for buffers and images.
Fixed a compilation bug in Mat4.hpp
Fixed a bug in GlPipeline, that produced an assertion in Signal.
Reworked debug logs.

+++ source/Renderer/VkRenderer/Src/Core/VkRenderingResources.cpp
@@ -17,7 +17,7 @@ namespace vk_renderer
 		m_fence = std::make_unique< Fence >( m_device, renderer::FenceCreateFlag::eSignaled );
 	}
 
-	bool RenderingResources::waitRecord( uint32_t timeout )
+	bool RenderingResources::waitRecord( uint64_t timeout )
 	{
 		bool res = m_fence->wait( timeout ) == renderer::WaitResult::eSuccess;
